Split build_yaml
Moved rule and platform functionality to separate files
