+++ ssg/build_yaml.py
@@ -12,17 +12,12 @@ import sys
 from xml.sax.saxutils import escape
 import glob
 
-import yaml
-
 import ssg.build_remediations
-from .build_cpe import CPEDoesNotExist, CPEALLogicalTest, CPEALFactRef
+from .build_cpe import CPEDoesNotExist
 from .constants import (XCCDF_REFINABLE_PROPERTIES,
                         SCE_SYSTEM,
-                        cce_uri,
                         dc_namespace,
-                        ocil_cs,
                         ocil_namespace,
-                        oval_namespace,
                         xhtml_namespace,
                         xsi_namespace,
                         stig_ns,
@@ -31,59 +26,22 @@ from .constants import (XCCDF_REFINABLE_PROPERTIES,
                         SSG_PROJECT_NAME,
                         SSG_REF_URIS,
                         PREFIX_TO_NS,
-                        FIX_TYPE_TO_SYSTEM
                         )
 from .rules import get_rule_dir_id, get_rule_dir_yaml, is_rule_dir
 from .rule_yaml import parse_prodtype
 
-from .cce import is_cce_format_valid, is_cce_value_valid
 from .yaml import DocumentationNotComplete, open_and_expand, open_and_macro_expand
 from .utils import required_key, mkdir_p
 
-from .xml import ElementTree as ET, add_xhtml_namespace, register_namespaces, parse_file
-from .shims import unicode_func
+from .xml import ElementTree as ET, register_namespaces, parse_file
 from .build_cpe import ProductCPEs
 import ssg.build_stig
 
-def dump_yaml_preferably_in_original_order(dictionary, file_object):
-    try:
-        return yaml.dump(dictionary, file_object, indent=4, sort_keys=False)
-    except TypeError as exc:
-        # Older versions of libyaml don't understand the sort_keys kwarg
-        if "sort_keys" not in str(exc):
-            raise exc
-        return yaml.dump(dictionary, file_object, indent=4)
 
-
-def add_sub_element(parent, tag, data):
-    """
-    Creates a new child element under parent with tag tag, and sets
-    data as the content under the tag. In particular, data is a string
-    to be parsed as an XML tree, allowing sub-elements of children to be
-    added.
-
-    If data should not be parsed as an XML tree, either escape the contents
-    before passing into this function, or use ElementTree.SubElement().
-
-    Returns the newly created subelement of type tag.
-    """
-    namespaced_data = add_xhtml_namespace(data)
-    # This is used because our YAML data contain XML and XHTML elements
-    # ET.SubElement() escapes the < > characters by &lt; and &gt;
-    # and therefore it does not add child elements
-    # we need to do a hack instead
-    # TODO: Remove this function after we move to Markdown everywhere in SSG
-    ustr = unicode_func('<{0} xmlns:xhtml="{2}">{1}</{0}>').format(tag, namespaced_data, xhtml_namespace)
-
-    try:
-        element = ET.fromstring(ustr.encode("utf-8"))
-    except Exception:
-        msg = ("Error adding subelement to an element '{0}' from string: '{1}'"
-               .format(parent.tag, ustr))
-        raise RuntimeError(msg)
-
-    parent.append(element)
-    return element
+from .entities.rule import Rule
+from .entities.common import XCCDFEntity, add_sub_element, add_warning_elements, add_nondata_subelements
+from .entities.platform import Platform, add_platform_if_not_defined
+from .entities import common
 
 
 def reorder_according_to_ordering(unordered, ordering, regex=None):
@@ -102,63 +60,6 @@ def reorder_according_to_ordering(unordered, ordering, regex=None):
                 unordered.remove(item)
     ordered.extend(sorted(unordered))
     return ordered
-
-
-def add_warning_elements(element, warnings):
-    # The use of [{dict}, {dict}] in warnings is to handle the following
-    # scenario where multiple warnings have the same category which is
-    # valid in SCAP and our content:
-    #
-    # warnings:
-    #     - general: Some general warning
-    #     - general: Some other general warning
-    #     - general: |-
-    #         Some really long multiline general warning
-    #
-    # Each of the {dict} should have only one key/value pair.
-    for warning_dict in warnings:
-        warning = add_sub_element(element, "warning", list(warning_dict.values())[0])
-        warning.set("category", list(warning_dict.keys())[0])
-
-
-def add_nondata_subelements(element, subelement, attribute, attr_data):
-    """Add multiple iterations of a sublement that contains an attribute but no data
-       For example, <requires id="my_required_id"/>"""
-    for data in attr_data:
-        req = ET.SubElement(element, subelement)
-        req.set(attribute, data)
-
-
-def check_warnings(xccdf_structure):
-    for warning_list in xccdf_structure.warnings:
-        if len(warning_list) != 1:
-            msg = "Only one key/value pair should exist for each warnings dictionary"
-            raise ValueError(msg)
-
-
-def add_reference_elements(element, references, ref_uri_dict):
-    for ref_type, ref_vals in references.items():
-        for ref_val in ref_vals.split(","):
-            # This assumes that a single srg key may have items from multiple SRG types
-            if ref_type == 'srg':
-                if ref_val.startswith('SRG-OS-'):
-                    ref_href = ref_uri_dict['os-srg']
-                elif ref_val.startswith('SRG-APP-'):
-                    ref_href = ref_uri_dict['app-srg']
-                else:
-                    raise ValueError("SRG {0} doesn't have a URI defined.".format(ref_val))
-            else:
-                try:
-                    ref_href = ref_uri_dict[ref_type]
-                except KeyError as exc:
-                    msg = (
-                        "Error processing reference {0}: {1} in Rule {2}."
-                        .format(ref_type, ref_vals, self.id_))
-                    raise ValueError(msg)
-
-            ref = ET.SubElement(element, 'reference')
-            ref.set("href", ref_href)
-            ref.text = ref_val
 
 
 def add_benchmark_metadata(element, contributors_file):
@@ -252,166 +153,6 @@ class SelectionHandler(object):
         updated_refinements = self._subtract_refinements(extended_refinements)
         updated_refinements.update(self.refine_rules)
         self.refine_rules = updated_refinements
-
-
-class XCCDFEntity(object):
-    """
-    This class can load itself from a YAML with Jinja macros,
-    and it can also save itself to YAML.
-
-    It is supposed to work with the content in the project,
-    when entities are defined in the benchmark tree,
-    and they are compiled into flat YAMLs to the build directory.
-    """
-    KEYS = dict(
-            id_=lambda: "",
-            definition_location=lambda: "",
-    )
-
-    MANDATORY_KEYS = set()
-
-    GENERIC_FILENAME = ""
-    ID_LABEL = "id"
-
-    def __init__(self, id_):
-        super(XCCDFEntity, self).__init__()
-        self._assign_defaults()
-        self.id_ = id_
-
-    def _assign_defaults(self):
-        for key, default in self.KEYS.items():
-            default_val = default()
-            if isinstance(default_val, RuntimeError):
-                default_val = None
-            setattr(self, key, default_val)
-
-    @classmethod
-    def get_instance_from_full_dict(cls, data):
-        """
-        Given a defining dictionary, produce an instance
-        by treating all dict elements as attributes.
-
-        Extend this if you want tight control over the instance creation process.
-        """
-        entity = cls(data["id_"])
-        for key, value in data.items():
-            setattr(entity, key, value)
-        return entity
-
-    @classmethod
-    def process_input_dict(cls, input_contents, env_yaml, product_cpes=None):
-        """
-        Take the contents of the definition as a dictionary, and
-        add defaults or raise errors if a required member is not present.
-
-        Extend this if you want to add, remove or alter the result
-        that will constitute the new instance.
-        """
-        data = dict()
-
-        for key, default in cls.KEYS.items():
-            if key in input_contents:
-                if input_contents[key] is not None:
-                    data[key] = input_contents[key]
-                del input_contents[key]
-                continue
-
-            if key not in cls.MANDATORY_KEYS:
-                data[key] = cls.KEYS[key]()
-            else:
-                msg = (
-                    "Key '{key}' is mandatory for definition of '{class_name}'."
-                    .format(key=key, class_name=cls.__name__))
-                raise ValueError(msg)
-
-        return data
-
-    @classmethod
-    def parse_yaml_into_processed_dict(cls, yaml_file, env_yaml=None, product_cpes=None):
-        """
-        Given yaml filename and environment info, produce a dictionary
-        that defines the instance to be created.
-        This wraps :meth:`process_input_dict` and it adds generic keys on the top:
-
-        - `id_` as the entity ID that is deduced either from thefilename,
-          or from the parent directory name.
-        - `definition_location` as the original location whenre the entity got defined.
-        """
-        file_basename = os.path.basename(yaml_file)
-        entity_id = file_basename.split(".")[0]
-        if file_basename == cls.GENERIC_FILENAME:
-            entity_id = os.path.basename(os.path.dirname(yaml_file))
-
-        if env_yaml:
-            env_yaml[cls.ID_LABEL] = entity_id
-        yaml_data = open_and_macro_expand(yaml_file, env_yaml)
-
-        try:
-            processed_data = cls.process_input_dict(yaml_data, env_yaml, product_cpes)
-        except ValueError as exc:
-            msg = (
-                "Error processing {yaml_file}: {exc}"
-                .format(yaml_file=yaml_file, exc=str(exc)))
-            raise ValueError(msg)
-
-        if yaml_data:
-            msg = (
-                "Unparsed YAML data in '{yaml_file}': {keys}"
-                .format(yaml_file=yaml_file, keys=list(yaml_data.keys())))
-            raise RuntimeError(msg)
-
-        if not processed_data.get("definition_location", ""):
-            processed_data["definition_location"] = yaml_file
-
-        processed_data["id_"] = entity_id
-
-        return processed_data
-
-    @classmethod
-    def from_yaml(cls, yaml_file, env_yaml=None, product_cpes=None):
-        yaml_file = os.path.normpath(yaml_file)
-
-        local_env_yaml = None
-        if env_yaml:
-            local_env_yaml = dict()
-            local_env_yaml.update(env_yaml)
-
-        try:
-            data_dict = cls.parse_yaml_into_processed_dict(yaml_file, local_env_yaml, product_cpes)
-        except DocumentationNotComplete as exc:
-            raise
-        except Exception as exc:
-            msg = (
-                "Error loading a {class_name} from {filename}: {error}"
-                .format(class_name=cls.__name__, filename=yaml_file, error=str(exc)))
-            raise RuntimeError(msg)
-
-        result = cls.get_instance_from_full_dict(data_dict)
-
-        return result
-
-    def represent_as_dict(self):
-        """
-        Produce a dict representation of the class.
-
-        Extend this method if you need the representation to be different from the object.
-        """
-        data = dict()
-        for key in self.KEYS:
-            value = getattr(self, key)
-            if value or True:
-                data[key] = getattr(self, key)
-        del data["id_"]
-        return data
-
-    def dump_yaml(self, file_name, documentation_complete=True):
-        to_dump = self.represent_as_dict()
-        to_dump["documentation_complete"] = documentation_complete
-        with open(file_name, "w+") as f:
-            dump_yaml_preferably_in_original_order(to_dump, f)
-
-    def to_xml_element(self):
-        raise NotImplementedError()
 
 
 class Profile(XCCDFEntity, SelectionHandler):
@@ -758,7 +499,7 @@ class Value(XCCDFEntity):
     def from_yaml(cls, yaml_file, env_yaml=None, product_cpes=None):
         value = super(Value, cls).from_yaml(yaml_file, env_yaml)
 
-        check_warnings(value)
+        common.check_warnings(value)
 
         return value
 
@@ -1248,510 +989,6 @@ def rule_filter_from_def(filterdef):
     return filterfunc
 
 
-class Rule(XCCDFEntity):
-    """Represents XCCDF Rule
-    """
-    KEYS = dict(
-        prodtype=lambda: "all",
-        title=lambda: "",
-        description=lambda: "",
-        rationale=lambda: "",
-        severity=lambda: "",
-        references=lambda: dict(),
-        identifiers=lambda: dict(),
-        ocil_clause=lambda: None,
-        ocil=lambda: None,
-        oval_external_content=lambda: None,
-        fixtext=lambda: "",
-        srg_requirement=lambda: "",
-        warnings=lambda: list(),
-        conflicts=lambda: list(),
-        requires=lambda: list(),
-        platform=lambda: None,
-        platforms=lambda: set(),
-        sce_metadata=lambda: dict(),
-        inherited_platforms=lambda: list(),
-        template=lambda: None,
-        cpe_platform_names=lambda: set(),
-        inherited_cpe_platform_names=lambda: list(),
-        bash_conditional=lambda: None,
-        fixes=lambda: dict(),
-        ** XCCDFEntity.KEYS
-    )
-
-    MANDATORY_KEYS = {
-        "title",
-        "description",
-        "rationale",
-        "severity",
-    }
-
-    GENERIC_FILENAME = "rule.yml"
-    ID_LABEL = "rule_id"
-
-    PRODUCT_REFERENCES = ("stigid", "cis",)
-    GLOBAL_REFERENCES = ("srg", "vmmsrg", "disa", "cis-csc",)
-
-    def __init__(self, id_):
-        super(Rule, self).__init__(id_)
-        self.sce_metadata = None
-
-    def __deepcopy__(self, memo):
-        cls = self.__class__
-        result = cls.__new__(cls)
-        memo[id(self)] = result
-        for k, v in self.__dict__.items():
-            # These are difficult to deep copy, so let's just re-use them.
-            if k != "template" and k != "local_env_yaml":
-                setattr(result, k, deepcopy(v, memo))
-            else:
-                setattr(result, k, v)
-        return result
-
-    @classmethod
-    def from_yaml(cls, yaml_file, env_yaml=None, product_cpes=None, sce_metadata=None):
-        rule = super(Rule, cls).from_yaml(yaml_file, env_yaml, product_cpes)
-
-        # platforms are read as list from the yaml file
-        # we need them to convert to set again
-        rule.platforms = set(rule.platforms)
-
-        # rule.platforms.update(set(rule.inherited_platforms))
-
-        check_warnings(rule)
-
-        # ensure that content of rule.platform is in rule.platforms as
-        # well
-        if rule.platform is not None:
-            rule.platforms.add(rule.platform)
-
-        # Convert the platform names to CPE names
-        # But only do it if an env_yaml was specified (otherwise there would be no product CPEs
-        # to lookup), and the rule's prodtype matches the product being built
-        # also if the rule already has cpe_platform_names specified (compiled rule)
-        # do not evaluate platforms again
-        if (
-                env_yaml and env_yaml["product"] in parse_prodtype(rule.prodtype)
-                or env_yaml and rule.prodtype == "all") and product_cpes and not rule.cpe_platform_names:
-            # parse platform definition and get CPEAL platform
-            for platform in rule.platforms:
-                cpe_platform = Platform.from_text(platform, product_cpes)
-                cpe_platform = add_platform_if_not_defined(cpe_platform, product_cpes)
-                rule.cpe_platform_names.add(cpe_platform.id_)
-
-
-        if sce_metadata and rule.id_ in sce_metadata:
-            rule.sce_metadata = sce_metadata[rule.id_]
-            rule.sce_metadata["relative_path"] = os.path.join(
-                env_yaml["product"], "checks/sce", rule.sce_metadata['filename'])
-
-        rule.validate_prodtype(yaml_file)
-        rule.validate_identifiers(yaml_file)
-        rule.validate_references(yaml_file)
-        return rule
-
-    def _verify_stigid_format(self, product):
-        stig_id = self.references.get("stigid", None)
-        if not stig_id:
-            return
-        if "," in stig_id:
-            raise ValueError("Rules can not have multiple STIG IDs.")
-
-    def _verify_disa_cci_format(self):
-        cci_id = self.references.get("disa", None)
-        if not cci_id:
-            return
-        cci_ex = re.compile(r'^CCI-[0-9]{6}$')
-        for cci in cci_id.split(","):
-            if not cci_ex.match(cci):
-                raise ValueError("CCI '{}' is in the wrong format! "
-                                 "Format should be similar to: "
-                                 "CCI-XXXXXX".format(cci))
-        self.references["disa"] = cci_id
-
-    def normalize(self, product):
-        try:
-            self.make_refs_and_identifiers_product_specific(product)
-            self.make_template_product_specific(product)
-        except Exception as exc:
-            msg = (
-                "Error normalizing '{rule}': {msg}"
-                .format(rule=self.id_, msg=str(exc))
-            )
-            raise RuntimeError(msg)
-
-    def add_stig_references(self, stig_references):
-        stig_id = self.references.get("stigid", None)
-        if not stig_id:
-            return
-        reference = stig_references.get(stig_id, None)
-        if not reference:
-            return
-        self.references["stigref"] = reference
-
-    def _get_product_only_references(self):
-        product_references = dict()
-
-        for ref in Rule.PRODUCT_REFERENCES:
-            start = "{0}@".format(ref)
-            for gref, gval in self.references.items():
-                if ref == gref or gref.startswith(start):
-                    product_references[gref] = gval
-        return product_references
-
-    def make_template_product_specific(self, product):
-        product_suffix = "@{0}".format(product)
-
-        if not self.template:
-            return
-
-        not_specific_vars = self.template.get("vars", dict())
-        specific_vars = self._make_items_product_specific(
-            not_specific_vars, product_suffix, True)
-        self.template["vars"] = specific_vars
-
-        not_specific_backends = self.template.get("backends", dict())
-        specific_backends = self._make_items_product_specific(
-            not_specific_backends, product_suffix, True)
-        self.template["backends"] = specific_backends
-
-    def make_refs_and_identifiers_product_specific(self, product):
-        product_suffix = "@{0}".format(product)
-
-        product_references = self._get_product_only_references()
-        general_references = self.references.copy()
-        for todel in product_references:
-            general_references.pop(todel)
-        for ref in Rule.PRODUCT_REFERENCES:
-            if ref in general_references:
-                msg = "Unexpected reference identifier ({0}) without "
-                msg += "product qualifier ({0}@{1}) while building rule "
-                msg += "{2}"
-                msg = msg.format(ref, product, self.id_)
-                raise ValueError(msg)
-
-        to_set = dict(
-            identifiers=(self.identifiers, False),
-            general_references=(general_references, True),
-            product_references=(product_references, False),
-        )
-        for name, (dic, allow_overwrites) in to_set.items():
-            try:
-                new_items = self._make_items_product_specific(
-                    dic, product_suffix, allow_overwrites)
-            except ValueError as exc:
-                msg = (
-                    "Error processing {what} for rule '{rid}': {msg}"
-                    .format(what=name, rid=self.id_, msg=str(exc))
-                )
-                raise ValueError(msg)
-            dic.clear()
-            dic.update(new_items)
-
-        self.references = general_references
-        self._verify_disa_cci_format()
-        self.references.update(product_references)
-
-        self._verify_stigid_format(product)
-
-    def _make_items_product_specific(self, items_dict, product_suffix, allow_overwrites=False):
-        new_items = dict()
-        for full_label, value in items_dict.items():
-            if "@" not in full_label and full_label not in new_items:
-                new_items[full_label] = value
-                continue
-
-            label = full_label.split("@")[0]
-
-            # this test should occur before matching product_suffix with the product qualifier
-            # present in the reference, so it catches problems even for products that are not
-            # being built at the moment
-            if label in Rule.GLOBAL_REFERENCES:
-                msg = (
-                    "You cannot use product-qualified for the '{item_u}' reference. "
-                    "Please remove the product-qualifier and merge values with the "
-                    "existing reference if there is any. Original line: {item_q}: {value_q}"
-                    .format(item_u=label, item_q=full_label, value_q=value)
-                )
-                raise ValueError(msg)
-
-            if not full_label.endswith(product_suffix):
-                continue
-
-            if label in items_dict and not allow_overwrites and value != items_dict[label]:
-                msg = (
-                    "There is a product-qualified '{item_q}' item, "
-                    "but also an unqualified '{item_u}' item "
-                    "and those two differ in value - "
-                    "'{value_q}' vs '{value_u}' respectively."
-                    .format(item_q=full_label, item_u=label,
-                            value_q=value, value_u=items_dict[label])
-                )
-                raise ValueError(msg)
-            new_items[label] = value
-        return new_items
-
-    def validate_identifiers(self, yaml_file):
-        if self.identifiers is None:
-            raise ValueError("Empty identifier section in file %s" % yaml_file)
-
-        # Validate all identifiers are non-empty:
-        for ident_type, ident_val in self.identifiers.items():
-            if not isinstance(ident_type, str) or not isinstance(ident_val, str):
-                raise ValueError("Identifiers and values must be strings: %s in file %s"
-                                 % (ident_type, yaml_file))
-            if ident_val.strip() == "":
-                raise ValueError("Identifiers must not be empty: %s in file %s"
-                                 % (ident_type, yaml_file))
-            if ident_type[0:3] == 'cce':
-                if not is_cce_format_valid(ident_val):
-                    raise ValueError("CCE Identifier format must be valid: invalid format '%s' for CEE '%s'"
-                                     " in file '%s'" % (ident_val, ident_type, yaml_file))
-                if not is_cce_value_valid("CCE-" + ident_val):
-                    raise ValueError("CCE Identifier value is not a valid checksum: invalid value '%s' for CEE '%s'"
-                                     " in file '%s'" % (ident_val, ident_type, yaml_file))
-
-    def validate_references(self, yaml_file):
-        if self.references is None:
-            raise ValueError("Empty references section in file %s" % yaml_file)
-
-        for ref_type, ref_val in self.references.items():
-            if not isinstance(ref_type, str) or not isinstance(ref_val, str):
-                raise ValueError("References and values must be strings: %s in file %s"
-                                 % (ref_type, yaml_file))
-            if ref_val.strip() == "":
-                raise ValueError("References must not be empty: %s in file %s"
-                                 % (ref_type, yaml_file))
-
-        for ref_type, ref_val in self.references.items():
-            for ref in ref_val.split(","):
-                if ref.strip() != ref:
-                    msg = (
-                        "Comma-separated '{ref_type}' reference "
-                        "in {yaml_file} contains whitespace."
-                        .format(ref_type=ref_type, yaml_file=yaml_file))
-                    raise ValueError(msg)
-
-    def validate_prodtype(self, yaml_file):
-        for ptype in self.prodtype.split(","):
-            if ptype.strip() != ptype:
-                msg = (
-                    "Comma-separated '{prodtype}' prodtype "
-                    "in {yaml_file} contains whitespace."
-                    .format(prodtype=self.prodtype, yaml_file=yaml_file))
-                raise ValueError(msg)
-
-    def add_fixes(self, fixes):
-        self.fixes = fixes
-
-    def _add_fixes_elements(self, rule_el):
-        for fix_type, fix in self.fixes.items():
-            fix_el = ET.SubElement(rule_el, "fix")
-            fix_el.set("system", FIX_TYPE_TO_SYSTEM[fix_type])
-            fix_el.set("id", self.id_)
-            fix_contents, config = fix
-            for key in ssg.build_remediations.REMEDIATION_ELM_KEYS:
-                if config[key]:
-                    fix_el.set(key, config[key])
-            fix_el.text = fix_contents + "\n"
-            # Expand shell variables and remediation functions
-            # into corresponding XCCDF <sub> elements
-            ssg.build_remediations.expand_xccdf_subs(fix_el, fix_type)
-
-    def to_xml_element(self, env_yaml=None):
-        rule = ET.Element('Rule')
-        rule.set('selected', 'false')
-        rule.set('id', self.id_)
-        rule.set('severity', self.severity)
-        add_sub_element(rule, 'title', self.title)
-        add_sub_element(rule, 'description', self.description)
-        add_warning_elements(rule, self.warnings)
-
-        if env_yaml:
-            ref_uri_dict = env_yaml['reference_uris']
-        else:
-            ref_uri_dict = SSG_REF_URIS
-        add_reference_elements(rule, self.references, ref_uri_dict)
-
-        add_sub_element(rule, 'rationale', self.rationale)
-
-        for cpe_platform_name in sorted(self.cpe_platform_names):
-            platform_el = ET.SubElement(rule, "platform")
-            platform_el.set("idref", "#"+cpe_platform_name)
-
-        add_nondata_subelements(rule, "requires", "idref", self.requires)
-        add_nondata_subelements(rule, "conflicts", "idref", self.conflicts)
-
-        for ident_type, ident_val in self.identifiers.items():
-            ident = ET.SubElement(rule, 'ident')
-            if ident_type == 'cce':
-                ident.set('system', cce_uri)
-                ident.text = ident_val
-        self._add_fixes_elements(rule)
-
-        ocil_parent = rule
-        check_parent = rule
-
-        if self.sce_metadata:
-            # TODO: This is pretty much another hack, just like the previous OVAL
-            # one. However, we avoided the external SCE content as I'm not sure it
-            # is generally useful (unlike say, CVE checking with external OVAL)
-            #
-            # Additionally, we build the content (check subelement) here rather
-            # than in xslt due to the nature of our SCE metadata.
-            #
-            # Finally, before we begin, we might have an element with both SCE
-            # and OVAL. We have no way of knowing (right here) whether that is
-            # the case (due to a variety of issues, most notably, that linking
-            # hasn't yet occurred). So we must rely on the content author's
-            # good will, by annotating SCE content with a complex-check tag
-            # if necessary.
-
-            if 'complex-check' in self.sce_metadata:
-                # Here we have an issue: XCCDF allows EITHER one or more check
-                # elements OR a single complex-check. While we have an explicit
-                # case handling the OVAL-and-SCE interaction, OCIL entries have
-                # (historically) been alongside OVAL content and been in an
-                # "OR" manner -- preferring OVAL to SCE. In order to accomplish
-                # this, we thus need to add _yet another parent_ when OCIL data
-                # is present, and add update ocil_parent accordingly.
-                if self.ocil or self.ocil_clause:
-                    ocil_parent = ET.SubElement(ocil_parent, "complex-check")
-                    ocil_parent.set('operator', 'OR')
-
-                check_parent = ET.SubElement(ocil_parent, "complex-check")
-                check_parent.set('operator', self.sce_metadata['complex-check'])
-
-            # Now, add the SCE check element to the tree.
-            check = ET.SubElement(check_parent, "check")
-            check.set("system", SCE_SYSTEM)
-
-            if 'check-import' in self.sce_metadata:
-                if isinstance(self.sce_metadata['check-import'], str):
-                    self.sce_metadata['check-import'] = [self.sce_metadata['check-import']]
-                for entry in self.sce_metadata['check-import']:
-                    check_import = ET.SubElement(check, 'check-import')
-                    check_import.set('import-name', entry)
-                    check_import.text = None
-
-            if 'check-export' in self.sce_metadata:
-                if isinstance(self.sce_metadata['check-export'], str):
-                    self.sce_metadata['check-export'] = [self.sce_metadata['check-export']]
-                for entry in self.sce_metadata['check-export']:
-                    export, value = entry.split('=')
-                    check_export = ET.SubElement(check, 'check-export')
-                    check_export.set('value-id', value)
-                    check_export.set('export-name', export)
-                    check_export.text = None
-
-            check_ref = ET.SubElement(check, "check-content-ref")
-            href = self.sce_metadata['relative_path']
-            check_ref.set("href", href)
-
-        check = ET.SubElement(check_parent, 'check')
-        check.set("system", oval_namespace)
-        check_content_ref = ET.SubElement(check, "check-content-ref")
-        if self.oval_external_content:
-            check_content_ref.set("href", self.oval_external_content)
-        else:
-            # TODO: This is pretty much a hack, oval ID will be the same as rule ID
-            #       and we don't want the developers to have to keep them in sync.
-            #       Therefore let's just add an OVAL ref of that ID.
-            # TODO  Can we not add the check element if the rule doesn't have an OVAL check?
-            #       At the moment, the check elements of rules without OVAL are removed by
-            #       relabel_ids.py
-            check_content_ref.set("href", "oval-unlinked.xml")
-            check_content_ref.set("name", self.id_)
-
-        if self.ocil or self.ocil_clause:
-            ocil_check = ET.SubElement(check_parent, "check")
-            ocil_check.set("system", ocil_cs)
-            ocil_check_ref = ET.SubElement(ocil_check, "check-content-ref")
-            ocil_check_ref.set("href", "ocil-unlinked.xml")
-            ocil_check_ref.set("name", self.id_ + "_ocil")
-
-        return rule
-
-    def to_file(self, file_name):
-        root = self.to_xml_element()
-        tree = ET.ElementTree(root)
-        tree.write(file_name)
-
-    def to_ocil(self):
-        if not self.ocil and not self.ocil_clause:
-            raise ValueError("Rule {0} doesn't have OCIL".format(self.id_))
-        # Create <questionnaire> for the rule
-        questionnaire = ET.Element("questionnaire", id=self.id_ + "_ocil")
-        title = ET.SubElement(questionnaire, "title")
-        title.text = self.title
-        actions = ET.SubElement(questionnaire, "actions")
-        test_action_ref = ET.SubElement(actions, "test_action_ref")
-        test_action_ref.text = self.id_ + "_action"
-        # Create <boolean_question_test_action> for the rule
-        action = ET.Element(
-            "boolean_question_test_action",
-            id=self.id_ + "_action",
-            question_ref=self.id_ + "_question")
-        when_true = ET.SubElement(action, "when_true")
-        result = ET.SubElement(when_true, "result")
-        result.text = "PASS"
-        when_true = ET.SubElement(action, "when_false")
-        result = ET.SubElement(when_true, "result")
-        result.text = "FAIL"
-        # Create <boolean_question>
-        boolean_question = ET.Element(
-            "boolean_question", id=self.id_ + "_question")
-        # TODO: The contents of <question_text> element used to be broken in
-        # the legacy XSLT implementation. The following code contains hacks
-        # to get the same results as in the legacy XSLT implementation.
-        # This enabled us a smooth transition to new OCIL generator
-        # without a need to mass-edit rule YAML files.
-        # We need to solve:
-        # TODO: using variables (aka XCCDF Values) in OCIL content
-        # TODO: using HTML formating tags eg. <pre> in OCIL content
-        #
-        # The "ocil" key in compiled rules contains HTML and XML elements
-        # but OCIL question texts shouldn't contain HTML or XML elements,
-        # therefore removing them.
-        if self.ocil is not None:
-            ocil_without_tags = re.sub(r"</?[^>]+>", "", self.ocil)
-        else:
-            ocil_without_tags = ""
-        # The "ocil" key in compiled rules contains XML entities which would
-        # be escaped by ET.Subelement() so we need to use add_sub_element()
-        # instead because we don't want to escape them.
-        question_text = add_sub_element(
-            boolean_question, "question_text", ocil_without_tags)
-        # The "ocil_clause" key in compiled rules also contains HTML and XML
-        # elements but unlike the "ocil" we want to escape the '<' and '>'
-        # characters.
-        # The empty ocil_clause causing broken question is in line with the
-        # legacy XSLT implementation.
-        ocil_clause = self.ocil_clause if self.ocil_clause else ""
-        question_text.text = (
-            u"{0}\n      Is it the case that {1}?\n      ".format(
-                question_text.text if question_text.text is not None else "",
-                ocil_clause))
-        return (questionnaire, action, boolean_question)
-
-    def __hash__(self):
-        """ Controls are meant to be unique, so using the
-        ID should suffice"""
-        return hash(self.id_)
-
-    def __eq__(self, other):
-        return isinstance(other, self.__class__) and self.id_ == other.id_
-
-    def __ne__(self, other):
-        return not self != other
-
-    def __lt__(self, other):
-        return self.id_ < other.id_
-
-    def __str__(self):
-        return self.id_
-
-
 class DirectoryLoader(object):
     def __init__(self, profiles_dir, env_yaml, product_cpes):
         self.benchmark_file = None
@@ -2059,94 +1296,3 @@ class LinearLoader(object):
             test_actions.append(action)
             questions.append(boolean_question)
         tree.write(filename)
-
-
-class Platform(XCCDFEntity):
-
-    KEYS = dict(
-        name=lambda: "",
-        original_expression=lambda: "",
-        xml_content=lambda: "",
-        bash_conditional=lambda: "",
-        ansible_conditional=lambda: "",
-        ** XCCDFEntity.KEYS
-    )
-
-    MANDATORY_KEYS = [
-        "name",
-        "xml_content",
-        "original_expression",
-        "bash_conditional",
-        "ansible_conditional"
-    ]
-
-    prefix = "cpe-lang"
-    ns = PREFIX_TO_NS[prefix]
-
-    @classmethod
-    def from_text(cls, expression, product_cpes):
-        if not product_cpes:
-            return None
-        test = product_cpes.algebra.parse(
-            expression, simplify=True)
-        id = test.as_id()
-        platform = cls(id)
-        platform.test = test
-        platform.test.enrich_with_cpe_info(product_cpes)
-        platform.name = id
-        platform.original_expression = expression
-        platform.xml_content = platform.get_xml()
-        platform.bash_conditional = platform.test.to_bash_conditional()
-        platform.ansible_conditional = platform.test.to_ansible_conditional()
-        return platform
-
-    def get_xml(self):
-        cpe_platform = ET.Element("{%s}platform" % Platform.ns)
-        cpe_platform.set('id', self.name)
-        # in case the platform contains only single CPE name, fake the logical test
-        # we have to athere to CPE specification
-        if isinstance(self.test, CPEALFactRef):
-            cpe_test = ET.Element("{%s}logical-test" % CPEALLogicalTest.ns)
-            cpe_test.set('operator', 'AND')
-            cpe_test.set('negate', 'false')
-            cpe_test.append(self.test.to_xml_element())
-            cpe_platform.append(cpe_test)
-        else:
-            cpe_platform.append(self.test.to_xml_element())
-        xmlstr = ET.tostring(cpe_platform).decode()
-        return xmlstr
-
-    def to_xml_element(self):
-        return self.xml_content
-
-    def to_bash_conditional(self):
-        return self.bash_conditional
-
-    def to_ansible_conditional(self):
-        return self.ansible_conditional
-
-    @classmethod
-    def from_yaml(cls, yaml_file, env_yaml=None, product_cpes=None):
-        platform = super(Platform, cls).from_yaml(yaml_file, env_yaml)
-        platform.xml_content = ET.fromstring(platform.xml_content)
-        # if we did receive a product_cpes, we can restore also the original test object
-        # it can be later used e.g. for comparison
-        if product_cpes:
-            platform.test = product_cpes.algebra.parse(
-                platform.original_expression, simplify=True)
-        return platform
-
-    def __eq__(self, other):
-        if not isinstance(other, Platform):
-            return False
-        else:
-            return self.test == other.test
-
-
-def add_platform_if_not_defined(platform, product_cpes):
-    # check if the platform is already in the dictionary. If yes, return the existing one
-    for p in product_cpes.platforms.values():
-        if platform == p:
-            return p
-    product_cpes.platforms[platform.id_] = platform
-    return platform
